Refactor menu page to display items in rows of three and update menu pages with Link components (#17)
This pull request includes several commits that refactor the menu pages to display items in rows of three and update the navigation buttons to use Link components instead of anchor tags. Additionally, revalidateTag calls were added to Base.js and Drive.js to ensure that cached data is updated when changes are made.